Don't make clean in the GIGC dir unless we compile with the hpc target
In this version, we modified the Makefile_header.mk so that we can
compile the standalone ESMF/MAPL GEOS-Chem with it.  But we neglected
to avoid trying to make clean in the GIGC subdirectory unless we
compile the hpc target.

Added an ifeq statement to the realclean target that will avoid this
from happening.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>