Correct floating-point exception in CHEM_SO2
In routine CHEM_SO2 in GeosCore/sulfate_mod.F, the scalar Ki was only defined
when using the offline aerosol simulation. The acid uptake by dust aerosol
code utilizes Ki when computing gas phase SO4 production. The GEOS-Chem
Unit Tester revealed that the undefined Ki was randomly assigned a negative
number, which caused the LOG10 call to choke.

The definitions for K0 and K1 have now been moved outside of the IF statement,
so that they are defined regardless of the simulation type. These
reaction rates were updated to the JPL 2006 values in v8-02-04 by Jenny
Fisher.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>