Removed SEAC4RS-specific modifications from the std code
(1) Removed the #if defined( SEAC4RS ) C-preprocesor switch.  This
was a holdover when the SEAC4RS grid was defined for SE Asia.  We now
use the NESTED_NA Cpp switch to run SEAC4RS simulations.

(2) Continued removing MEGAN 2.1 and NEI2008 updates from routines.
    Those will be brought in later, and will have to be benchmarked
    separately.

(3) "Parked" the following routines in the obsolete directory:

     (a) finn_biomass_mod.F
     (b) flambe_biomass_mod.F
     (c) ncdf_mod.F
     (d) nei2008_anthro_mod.F

     These are not really obsolete, but we need to store them somewhere
     until they are needed again.  NOTE: ncdf_mod.F was packaged with
     finn_biomass_mod.F, but we should really use the NcdfUtil code to
     read those emissions.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>