Update regridding algorithm in POPS_READYR to MAP_A2A
Routine POPS_READYR (in GeosCore/pops_mod.F) was still utilizing the old
regridding routine DO_REGRID_1x1. This code has now been updated to use the
MAP_A2A regridding algorithm introduced in v9-01-03.

The exisiting code in POPS_READYR was for data on the GMAO 1x1 grid, but the
POPs data files are on the generic 1x1 grid (360x180). The dimensions of
arrays and the calls to the read/regrid routines have been updated to use
dimensions I1x1, J1x1-1.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>