Prevent div-by-zero in GeosCore/paranox_mod.F
In GEOS-Chem v9-02, we discovered a div-by-zero condition in
GeosCore/paranox_mod.F.  We have to make sure that the array
holding J(NO2) not zero before we put it into the denominator.

Bring this fix into the UCX code, which is based on v9-01-03,
and accordingly, which did not have this fix.

Signed-off-by: Sebastian David Eastham <seastham@mit.edu>