Now treat DST2, DST3, and DST4 as coarse mode in wet scavenging
While resolving some of the conflicts in the acid uptake on dust aerosol
code (based on v8-01-01), it was brought to Duncan’s attention that we now
use different scavenging coefficients for fine and coarse aerosols. In
v9-01-03, we implemented Qiaoqiao’s updates for aerosol scavenging
efficiency. From this update, we calculate WASHFRAC for accumulation mode
and coarse mode aerosol using washout efficiency from Feng (2007). This
paper has efficiency for aerosol sizes 0.001-0.04 um, 0.04-2.5 um
(accumulation mode), 2.5-16 um (coarse mode) and 16-100 um. Based on those
numbers, Qiaoqiao used accumulation mode for DST1-DST3, and coarse mode for
DST4. Duncan has suggested that we revisit what we consider fine and coarse
mode for dust in the wet deposition code. He thinks it may be better to
treat DST1 as accumulation mode and DST2-DST4  as coarse mode.

In the code, the dust size bins are defined as:

    DST1: 0.11.0 um
    DST2: 1.01.8 um
    DST3: 1.83.0 um
    DST4: 3.06.0 um

Jeff Pierce wrote:

    The dust bin sizes you provide are radius, whereas the efficiencies
    diameter (weird that accumulation mode could go up to D=5 um... even
    2.5 um is big for accumulation mode).  Thus, we need to multiply the
    dust-bin sizes by 2.  In which case DST3 and DST4 are both coarse and
    DST2 sits in between (with a larger fraction in coarse).  I support
    Duncan's suggestion.

    We should stop the use of radius for aerosol size measurements.  It's a
    minority, and causes errors and confusion in cases such as this.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>