Update dependencies listed in the Makefiles
Derived type objects are now used to pass information about input options,
the chemistry state, and the meteorology fields. Therefore, many modules
no longer depend on logical_mod.F, tracer_mod.F, and dao_mod.F.
We have gone through the routines in the GeosCore, GeosApm, GeosUtil, and
Headers directories and made sure that the dependencies listed in the
Makefiles are up-to-date.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>