Now read T_FULLGRID only for offline sims (commit #1)
We have made modifications in the code to restrict the allocation and
use of the T_FULLGRID and AIRDEN_FULLGRID arrays to offline simulations.
This will

We added field Input_Opt%ITS_A_SPECIALTY_SIM, which is TRUE only if we
are running a specialty simulation.  In routine INIT_DAO, we only allocate
T_FULLGRID and AIRDEN_FULLGRID if Input_Opt%ITS_A_SPECIALTY_SIM = .TRUE.

We also had to pass the Input_Opt object to various routines in the
offline simulations to get this to work.

We also now read T_FULLGRID for MERRA simulations. We are continuing to
add the code modifications for GEOS-FP.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>