Prevent log(0) from occurring in function PULSING in soil NOx module
In function PULSING (in module GeosCore/soil_nox_mod.F), a numerical
floating-point error may occur in the expression LOG( DRYPERIOD ), if
DRYPERIOD is zero.  To prevent this from occurring, we now only do the
logarithm if DRYPERIOD > 0.  If DRYPERIOD = 0, we assign the entire
expression a default value (-53.6).

This issue was discovered with the GEOS-Chem Unit Tester while validating
the GEOS-4 4x5 full-chemistry simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>