Bug fix: add GEOS-4 parameters to Headers/smv_dimension_mod.F
The module file Headers/smv_dimension_mod.F contains size parameters
that are needed to dimension the various FAST-J arrays (formerly in
common blocks, now in module files).  This was a modification that we
had made for the Grid-Independent GEOS-Chem after v9-01-03.

We had neglected to add the proper vertical dimension parameters for
the GEOS-4 grid (30 reduced levels, 55 full vertical levels).  This
has now been corrected.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>