Use State_Met%U and State_Met%V in GEOS-4 section of a6_read_mod.F
We had erroneously passed State_Met%UWND and State_Met%VWND to the
subroutine READ_A6.  But these are not field names in the State_Met
object.  We should have used State_Met%U and State_Met%V instead.
This was noted in the comments but evidently the change was never made.
We have now corrected this.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>