Eliminate array temporaries in GEOS-4 convection modules #1
We have rewritten the calls to routines FVDAS_CONVECT, HACK_CONV, and
ARCONVTRAN (in GeosCore/convection_mod.F and GeosCore/fvdas_convect_mod.F)
so as to avoid creating array temporaries.  As of this commit, we have
eliminated array temporaries for the HKETA, HKBETA, ZMMU, ZMMD, ZMEU
met fields.

Array temporaries still exist for the STT and F (aka FRACIS) arguments,
so we will remove them in a future commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>