Now use an #ifdef to pick the year of tagged O3 archived P/L data
Now use an internal #define statement

   #define USE_THIS_O3_YEAR YYYY

(where YYYY = 2003, 2004, 2005, etc.) to select the year of archived
O3 prod/loss data.  This allows the user to customize the tagged O3
simulation depending on the year of archived data.

Also, if this #define statement is commented out, then the code will
default to the current simulation year.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>