Now re-activate the parallel DO loop in LINOZ_CHEM3
The !$OMP PARALLEL DO loop in LINOZ_CHEM3 was deactivated because it
previously caused an error that could not be traced.

Recent runs (tagO3 and fullchem) with the GEOS-Chem Unit Tester have
shown that the parallel loop now functions properly when activated.
This will also lead to a gain in scalability in long tagO3 or fullchem
simulations.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>