Now replace NaN's in MOISTQ field with zeroes.
In the GEOS-5 MOISTQ met fields, there have been shown to be NaN values
for certain days near the tropopause.  These NaN values can be safely
reset to zero.  Near the tropopause there should not be any contributions
from cloud or ice precipitation, and so MOISTQ should be zero there.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>