Clean up ND43 and ND45 diagnostics
The HR1_NO and HR2_NO scalars have been removed and are no longer read from
input.geos. They aren't needed because NO, NO2, and NO3 are now tracers and
are included in the ND45 diagnostic.

We have also removed LTO3 and CTO3 for the ND45 diagnostic. They aren't used
anymore because O3 is now a tracer.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>