Prevent divide-by-zero in GET_DUST_ALK
In routine GET_DUST_ALK in GeosCore/dust_mod.F, only compute KT1 if
GAMS or GAMN is > 0. GAMS and GAMN may be zero when the RH is low. When
this is the case, we will set the gas-to-particle rate constant KT1 to
zero to prevent a floating-point invalid situation.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>