Bug fix: Convert MERRA, GEOS-FP relative humidity to percent
Due to historical reasons, GEOS-Chem carries relative humidity (RH) from
the met fields in %.  RH in GEOS-5 met is unitless, but is converted
to % in the code.

RH in the MERRA and GEOS-FP met products likewise are unitless.  But we
had forgotten to do the same unit conversion to %, as we already do for
GEOS-5.  We have now restored these unit conversions in modules
GeosCore/merra_i6_mod.F and GeosCore/geosfp_read_mod.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>