Remove double-counting of crop NH3 emissions
Fabien Paulot informed us that the "natural NH3" emissions used in GEOS-Chem
come from GEIA and include crops. This leads to a double-counting of these
emissions when NEI08 emissions are used. We now overwrite "natural" emissions
by a version derived from the original GEIA files for natural soil, wild
animals, and oceans when we are within the NEI08 mask.

Signed-off-by: Jenny Fisher <jennyf@uow.edu.au>