Need to add Sajeev Philip's PBL fix to the GEOS-FP met fields.
Sajeev Philip added a fix for the PBL height, so that the PBL height
must be greater than a minimum mixing depth, defined as 700*USTAR.
This fix was applied in the G-C code to GEOS-5 and MERRA met fields,
but not to GEOS-FP.  Add it in here.

NOTE: Sajeev is working on a better online PBL calculation which may
supersede this in the future.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>