Bug fixes - SP2
Corrected a bug introduced with the RCP emissions module
whereby NH3 emissions were not correctly scaled if RCP
emissions were disabled.

Added the ability to fix methane surface mixing ratios
to the value formerly used throughout for all levels at
the given latitude, as suggested by UCX reviewer #2.

Finalized the addition of D71, a new diagnostic which
records the hourly maximum mixing ratio of a tracer to
allow direct comparison to (for example) O3 monitoring
stations.

Expanded diagnostic 64 (D64) to report direct, diffuse and
net flux independently. THe previous implementation summed
fluxes to determine total energy available for photolysis
rather than preserving directionality as is necessary for
calculation of a net downward flux.

Further groundwork for precipitation adjustment in geoeng.
scenarios laid down.

Fixed a benign issue where a redundant reset of the H2O
fields in the troposphere was only called when the DEVEL
flag was set. This is now called in both instances for
safety's sake, although the reset is likely unnecessary
given that it already occurs elsewhere.

This version of the code was used to produce the SP2 results.

Signed-off-by: Sebastian David Eastham <seastham@mit.edu>