Update to the wet scavenging scheme from Qiaoqiao Wang.
"The major effect of the updates is to remove more BCPO and OCPO
during convection (mainly distributed in low latitudes). Consequently,
less BCPI and OCPI is present in the atmosphere. There is also minor
effect of the updates which slightly decrease water-soluble aerosols
in the upper troposphere through homogeneous ice nucleation."

Signed-off-by: Sungshik Kim <kim68@fas.harvard.edu>