Update dust scaling factors for 2x2.5 and nested NA GEOS-FP simulations
The dust scaling factor for 2x2.5 GEOS-FP simulations has been calculated
based on a 1-year simulation for 08/2012-08/2013.

Scaling factors for the 0.25x0.3125 nested NA simulation have been updated
to use dust emissions from the 08/2012-08/2014 period in the calculation.
That way, the GEOS-FP dust scaling factors are all calculated for the same
time period.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>