Minor updates in ProTeX headers
Fixed a few minor issues in ProTeX headers (mostly outdated and/or
missing comments).  Also updated the version of doc/protex so that it
is consistent with the version used in the NCL4GC package.

We now only save out PDF output.  PDF files are in general much smaller
than the equivalent PS files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>