Reactivate call to ADD_STRAT_POX in tagged Ox simulation
Jintai Lin noted that the stratospheric P(Ox) produced by the Synoz boundary
condition should be passed to the Tagged Ox simulation source code in
tagged_ox_mod.F. This is only for runs which use the Synoz boundary condition.
If you are using Linoz, then this should be disabled.

The updates are as follows:

(1) In UPBDFLX_O3, we need to define J30S and J30N parameters for nested
    grid regions. This will allow Tagged Ox simulations for the nested grid
    regions to be conducted.

    #elif defined( GRID05x0666 )

    ! Lin_20111026
    #if   defined( NESTED_CH )
          INTEGER, PARAMETER   :: J30S = 1,  J30N = 83
    #elif   defined( NESTED_NA )
          INTEGER, PARAMETER   :: J30S = 1,  J30N = 41
    #elif   defined( NESTED_EU )
          INTEGER, PARAMETER   :: J30S = 1,  J30N = 1         ! add later
    #endif

(2) Further down in routine UPBDFLX_O3, we must call ADD_STRAT_POx. This
    will pass the P(Ox) produced by Synoz to the source code in
    tagged_ox_mod.F. This call is placed immediately after updating the
    STT tracer array:

            ! Store O3 flux in the proper tracer number
            STT(I,J,L,IDTO3) = STT(I,J,L,IDTO3) + PO3

            ! Store O3 flux for strat Ox tracer (Tagged Ox only)
            ! UPBDFLX_O3 and thus ADD_STRAT_POX is called only
            ! when Synoz is used (LLINOZ is FALSE) (jtl, hyl, dbj, 11/3/11)
            IF (Input_OPt%ITS_A_TAGOX_SIM ) THEN
               CALL ADD_STRAT_POX( I, J, L, PO3 )
            ENDIF

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>