Avoid array temporaries in TOMAS simulations, update #3
Rewrote the subroutine calls to NH4BULKTOBIN to avoid the creation of
array temporaries.  In most cases this was done by replacing:

       CALL NH4BULKTOBIN( MK(1:IBINS,SRTSO4), STT(I,J,L,IDTNH4),
      &                   TRANSFER(1:IBINS)  )

with:
       CALL NH4BULKTOBIN( MK(:,SRTSO4), STT(I,J,L,IDTNH4),
      &                  TRANSFER )

By explicitly stating the sub-slice (1:IBINS,SRTSO4), this causes the
compiler to create an array temporary.  Using (:,SRTSO4) instead allows
for a more efficient pointer slice to be passed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>