TOMAS bug fix: Add missing variables to !$OMP+PRIVATE statement
In routine AEROPHYS (in GeosCore/tomas_mod.F), we need to add the
following variables to the !$OMP+PRIVATE statement in the parallel
DO loop: TRACNUM, NH3_TO_NH4, SURF_AREA.  Adding this now causes
TOMAS to have identical sp vs. mp results when running with just
chemistry turned on.

NOTE: The ND60 diagnostic appears to be problematic.  There may be
other variables that have to be held private for the parallel loop.
For further testing we shall just turn ND60 off and worry about this
later.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>