Avoid parallelization error in drydep_mod.F for TOMAS simulations
Instead of holding A_RADI(K) and A_DEN(K) as !$OMP+PRIVATE in TOMAS
simulations (in the main DO loop in DEPVEL), we now save the particle
size and density values to PRIVATE variables DIAM and DEN.  We then
pass those as arguments to function DUST_SFCRSII.

This will also affect the DST1-DST4 tracers in non-TOMAS fullchem
simulations.  We will test to make sure that our modification does
not break anything.  At first glance it appears that it does not.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>