Now place ND37 diagnostic update in same loop as COMPUTE_F
This is a correction for the TOMAS bug fix.  We have now pulled
the call to routine COMPUTE_F out of the main tracer loop and into
its own separate OMP-Parallel do loop.  But we also need to place
the IF block where the ND37 diagnostic is computed into the same
DO loop.

COMPUTE_F returns the ISOL varaiable, which needs to be held PRIVATE
for the DO loop.  All references to ISOL have to be in the same
DO loop as COMPUTE_F.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>