Bug fix for ND37, ND38 diagnostics in convection_mod.F
The modification for TOMAS wetdep (to pull COMPUTE_F out of the tracer
loop and into its own loop) also required the ND37 diagnostic IF block
to be put into the same loop as COMPUTE_F.  This was done in the
previous commit.

Furthermore, because COMPUTE_F returns the value of ISOL, we must
and because ISOL is also used for the ND38 diagnostic in the tracer
loop, we must also save the values of ISOL in a 1-D vector.  This will
allow the values of ISOL to be passed from the first parallel loop
to the second.

This modification has been tested in the GEOS_Chem Unit Tester by
Bob Yantosca (04 Feb 2014) and it has yielded identical results for
geos5_4x5_fullchem, geos5_4x5_Hg, geos5_4x5_RnPbBe, geos5_4x5_soa
and geos5_4x5_soa_svpoa simulations.

The geos5_4x5_TOMAS40 simulation still yields differences between
sp & mp tests.  There may be further issues to diagnose & correct.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>