Add fixes to drydep_mod.F from v9-02r
In GEOS-Chem v9-02r, we added a few minor fixes to drydep_mod.F.  A
couple of variables needed to be declared !$OMP+PRIVATE.  We also
added the better test for depositing species in DEPVEL.

We had to add these fixes in now so that the UCX code would produce
identical results on single and multi processors.  This will facilitate
further debugging in the GC Unit Tester.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>