Add missing variables to !$OMP+PRIVATE statements in setemis.F anf sulfate_mod.F
Ship emissions were moved to setemis.F, but several variables were missing
from the !OMP+PRIVATE statements. This causes differences in GEOS-Chem
diagnostics when comparing sp and mp output.

We have also added an !$OMP+PRIVATE statement to a DO loop in subroutine
READ_ANTHRO_NH3 (sulfate_mod.F). This bug was fixed in GEOS-Chem v9-02, but
was not included in the UCX version of the code.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>