Make UCX parallel DO loops more efficient (#1)
In routine UCX_NOX, we now compute ZMID outside of the main parallel
DO loop.  ZMID is now stored as a 3-D array which is then referenced
within the main parallel DO loop.  This lets us to change the loop
order of the main parallel DO loop to L-J-I, which will ensure that
we step through the arrays in the same way they are laid out in
memory.  This should execute much more efficiently.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>