Make UCX parallel DO loops more efficient (#6)
In routine UCX_H2SO4PHOT (in module GeosCore/ucx_mod.F), we now
have reordered the DO loops from I-J-L to J-I-L.  This is more
efficient.

Ideally we would like to put the L-loop on the outside so that
this would be L-J-I, which is the most efficient.  Do that later.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>