Clean up use of State_Met in several modules
In modues were we reference fields from the State_Met derived type object
several times, we now use a pointer to point to that field. This makes the
code more readable and will help reduce the number of conflicts when
merging updates into v9-02 code. These modifications are purely cosmetic,
and it has been verified that they do not change GEOS-Chem output.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>