Now obtain online O3 for Fast-JX from State_Chm%Tracers
In the original UCX code, online O3 was saved to O3GRID in partition.F. In
v9-02, partitioning of NOx and Ox into family tracers was removed and their
individual constituents (including O3) were made tracers. Therefore, we
can now directly copy concentrations of O3 stored in State_Met%Tracers to
O3_CTM in routine FAST_JX. Printing out the values of O3_CTM in the
original UCX code and with this fix verifies that the units are consistent
(molec/cm2).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>