Restore KPP_SOLVE_ALWAYS=yes option to skip past nonconvergence
Compiling with KPP_SOLVE_ALWAYS=yes now causes KPP to loosen tolerances
if a non-convergent box is encountered. This allows the GEOS-Chem run to
continue, instead of stopping the run after two non-convergences. Bob Y.
added this option to the SEAC4RS branch of the code, but it never made it
into the standard code.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>