Prevent sea salt from being emitted over ice in TOMAS simulations
Jeff Pierce (jeffrey.pierce@colostate.edu) wrote:

The FOCEAN (fraction of box that is ocean parameter) in TOMAS seasalt
emissions didn't consider ice.  I've modified it to do things more like
the bulk emissions.  In GeosCore/seasalt_mod.F, at line 1954, there is
a line...

    FOCEAN   = 1d0 - State_Met%FRCLND(I,J)

I've updated this to be...

    IF ( IS_WATER( I, J, State_Met ) ) THEN
       FOCEAN   = 1d0 - State_Met%FRCLND(I,J)
    ELSE
       FOCEAN = 0.d0
    ENDIF

I couldn't figure out a way for FOCEAN to take into account the
fraction that is land and fraction that is ice, so I will just use
the IS_WATER to filter out boxes that are mostly ice.  The box scheme
is actually simpler and emits into the full box (or not) depending
on the logical variable returned by function IS_WATER.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>