Minor fix for arrays in PHOTO_JX
Changed the dimension of DDJ, OOJ, and T_CLIM from JXL1_+1 to JXL1_. The
corresponding arrays in SET_PROF only have L1_ (= JXL1_) levels, so the
last element of these arrays was set to a bogus value. The T_INPUT array
is given an extra level for compatibility, so we now set T_INPUT(L1_+1)
to T_CLIM(L1_).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>