Add GTMM_Hg=yes to allhg, libhg, libgtmm, exehg targets in Makefile
In the top-level Makefile, we now add the GTMM_Hg=yes flag to
each of the targets allhg, libhg, libgtmm, exehg.  If GEOS-Chem
is compiled at the top level of the code directory, then the
GTMM_Hg setting will propagate down to all lower levels of code.
This will link the GTMM mercury model library into the code when
the executable is created.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>