Change relative tolerance for UCX simulation
When running the UCX simulation, non-convergence messages would print to
the log file for a few boxes at levels 50 and above. After several days,
the run would stop with a non-convergence error when the solver fails
twice. Changing RTOL from 2.d-1 to 2.d-2 seems to fix this problem. We
apply this tolerance only when UCX=y is specified at compile time.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>