Now call INIT_WETSCAV when chemistry is on to avoid OOB errors
When running simulations with wet deposition turned off, an array-out-of-
bounds error would occur for H2O2s and SO2s because these arrays are also
used in routine CHEM_SO2. To fix this, we now call INIT_WETSCAV when
chemistry is turned on, so that H2O2s and SO2s are allocated. We have also
updated routine INIT_WETSCAV so that the rest of the subroutine is only
executed when wet deposition is turned on.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>