Reduce memory footprint in GeosCore/strat_chem_mod.F90
The following arrays in GeosCore/strat_chem_mod.F90 can be declared
as REAL*4 instead of REAL*8:

(1) PROD(:,:,:,:)
(2) LOSS(:,:,:,:)
(3) STRAT_OH(:,:,:)
(4) MINIT(:,:,:)
(5) SCHEM_TEND(:,:,:,:)

For the most part, these arrays hold values that are read from netCDF
files as REAL*4 data, or are computed from REAL*4 data.  This change
now allows us to run the 2x25 model in the smp8 computational queue at
Harvard.  (Previously, the 2x25 model would exit with a seg fault in
this queue.)

This update also required a modification to transfer_mod.F.  We now had
to split routine TRANSFER_3D into two routines: TRANSFER_3D_R4 (with
an output argument of REAL*4) and TRANSFER_3D_R8 (with an output
argument of REAL*8).  These routines are overloaded with a module
interface.  This allows us to use TRANSFER_3D on the PROD, LOSS,
and STRAT_OH arrays in strat_chem_mod.F90, because these are now
REAL*4 instead of REAL*8.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>