Now allocate wetdep arrays if convection is turned on
To prevent array-out-of-bounds errors when testing, we must allocate the
wetdep arrays (Vud, C_H2O, CLDLIQ, and CLDICE) when convection is on. These
arrays are all used in COMPUTE_F, which is called from the routines in
convection_mod.F. The computation of these arrays must also be done if
convection is on, in order to avoid a divide-by-zero error.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>