Make sure ptr_Q is received by an assumed shape array in routine FZPPM
In routine TPCORE_FVDAS (in GeosCore/tpcore_fvdas_mod.F90), we use
the ptr_Q pointer to grab a 3-D slice from the 4-D Q qrray, and pass
that to routine FZPPM.

In routine FZPPM, the Q variable is received into dummy argument QQ1.
Formerly, QQ1 was declared with explicit bounds.  But because we now
pass the ptr_Q pointer, we need to declare QQ1 as an assumed shape.
This means using (:,:,:) instead of using explicit dimension limits.
Doing this is more efficient of memory and reduces the time spent in
the TPCORE_FVDAS routine substantially.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>