Speed up GEOS-FP and MERRA convection routine
Added two fixes to routine DO_CONVECTION (in GeosCore/convection_mod.f)
for MERRA and GEOS-FP met fields:

(1) Change the parallel do loop to !$OMP+SCHEDULE( DYNAMIC ).  This
    results in better load balancing.

(2) Now test if ( ND14 > 0 ) and ( ND38 > 0 ) outside of the parallel
    DO loop, for better efficiency.

Testing with the TAU profiler revealed a reduction in time spent in
this routine by about 3.3 times with these fixes applied.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>