Add fixes to calcrate.F for better numerical stability w/ UCX
In GeosCore/calcrate.F, we have made the following updates to provide
better numerical stability when using the UCX strat-trop mechanism.

(1) Resized the PSEDUCTCONC array from (13,2) to (11,2).  We were not
    using the (12,:) and (13,:) slots of this array anymore.

(2) In a couple of places, set XSTCKF = 0d0 instead of TINY(0d0).
    TINY(0d0) is a number like 1e-312 and that can cause underflow.

(3) Modify an error trap for the N2O5 + H2O reaction.  Only allow the
    division to take place if the H2O concentration (stored in
    the EDUCTCONC variable at that location in the code) is greater
    than 1e-36.  We had previously used the IF_SAFE_DIV function as
    an error trap.  However, it is possible for the division to
    be allowed numerically, but the result to be large enough to
    choke the solver.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>