Add fix to prevent NaN for WASHFRAC in convection_mod.F
In routine DO_MERRA_CONVECTION (in module file GeosCore/wetscav_mod.F),
we have now placed a similar error check after the call to routine
WASHOUT as is already done in the wetdep module wetscav_mod.F.  This
should prevent NaN's from propagating throughout the code.  We will
need to benchmark this to see if it has any negative impacts on the
simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>