Avoid conflicting use of reaction flag 'F' in globchem.dat
Sebastian Eastham wrote:

   Looking at readchem.F, it seems like 'F' in SPECL(1) might be used for
   two conflicting purposes - once as a counter in NNADDF, which appears
   to be related to the channels relevant to NKHOROI and NKHOROJ, and then
   again as a counter for HAC reaction branching.

To avoid a conflict, the use of 'F' for the HOC2H4O --> HO2  + 2CH2O
reaction has been changed to 'FF'. This change has also been made in the
globchem.dat file for the dicarbonyls simulation that ships with GEOS-Chem.

We have also uncommented the code for EP photolysis in calcrate.F and
blocked it off with an IF statement, so that the code will now only be
executed if the dicarbonyls mechanism is used. A note has been included so
users of the dicarbonyls simulation know that the code needs to be updated
for FAST-JX v7.0.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>