Remove offending code in hcox_soilnox_mod.F90
In previous commit 107796d, we fixed a floating-point error bug in the
new HEMCO extension module for soil NOx emissions (hcox_soilnox_mod.F90).
But we had forgotten to remove the offending line of code, which was
still placed in front of the new code.  Therefore, the error could still
occur under some conditions.  This is now corrected.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>