Activate HEMCO diagnostics in hcoi_gc_diagn_mod.F90 (debugging only)
In hcoi_gc_diagn_mod.F90, we now turn on the HEMCO diagnostics when
we compile GEOS-Chem with DEBUG=yes.  This is a temporary fix to
allow unit testing to occur.  We shall add a more permanent fix later.

We also have added cosmetic changes in hcoi_gc_diagn_mod.F90 and
hco_driver_mod.F90.  We now use F90 free-format indentation, and have
updated the ProTeX headers as in a previous commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>