Remove directory_mod.F; now replaced by Input_Opt object (step #1)
We now pass Input_Opt to various routines so that we can use the
Input_Opt directory fields instead of relying on the obsolete module
file directory_mod.F.

NOTE: More work needs to be done before this is completed.  Most of
the instances of directory_mod.F were in the specialty simulations.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>