Replace directory_mod.F with fields from Input_Opt
To facilitate the Grid-Independent GEOS-Chem (and also the implementation
of HEMCO), we moved directory_mod.F to the obsolete directory.  Fields
from directory_mod.F are now supplied by Input_Opt.

NOTE: At present, we need a way to pass Input_Opt%DATA_DIR_1x1 to
the paranox_util_mod.F90.  This probably needs to be done through
the HEMCO ExtState object.  Still working on this.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>