Removed references to tracer_mod.F from many routines
Much of the functionality of tracer_mod.F is now obsolete with the
advent of the Input_Opt derived-type object.  We have now replaced
most of the fields and functions of tracer_mod.F with Input_Opt.

NOTE: We still have kept the INIT_TRACER routine (which is used to
initialize a shadow copy of Input_Opt%N_TRACERS within tracer_mod.F)
and the CHECK_STT* routines.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>