Prevent taking LOG(0) in ocean_toolbox_mod.F90
Added an error trap to prevent us from taking the natural logarithm
unless the argument is nonzero.  This is sometimes necessary in order
to pass the various unit tests.

We should understand better why the logarithm of zero occurs in the
first place.  This may be attributable to bad input.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>