Bug Fix: Restore DTCHEM in CHEM_TAGGED_CO
In a commit from Jan 2013 title "Replace STT with State_Chm' in v9-02, the
assignment of the variable DTCHEM was mistakenly removed from subroutine
CHEM_TAGGED_CO. DTCHEM is used all over the place in that subroutine, so
this resulted in various problems (unexpected zeros, infinities, etc.). I
assume that it was actually DTSRCE that was meant to be removed (since
DTSRCE is not used here), so I have removed that along with a few other
obsolete local variables. I have also rearranged to move parameter
definitions out of the main code.

Signed-off-by: Jenny Fisher <jennyf@uow.edu.au>