Remove inadvertently-added error check from convection_mod.F
We have removed an error check on WASHFRAC from convection routine
DO_MERRA_CONVECTION in GeosCore/convection_mod.F.  This had been
added in a development version of GEOS-Chem but was never added into
the standard benchmarked GEOS-Chem version.

For consistency with the previous version v10-01d, we have now
removed this from the code.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>