Avoid reading ASCII file for Ott et al 2010 lightning CDF's
We now hardwire the Ott et al CDF's that are used to partition
the column lightning NOx into the vertical.  This was formerly
read from ASCII file light_dist.ott2010.dat.

Hardwiring this data avoids several issues when running GEOS-Chem
in the ESMF environment, because (1) ASCII files need to be read
only on the root CPU (and the data broadcasted to other CPUs), and
(2) ESMF/MAPL can only regrid data that comes on a lon/lat grid.

Also made several other minor fixes and cosmetic changes to a few
other files.  Added .gitignore files in the HEMCO directories.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>