Remove boxvl.F; now compute it from State_Met%AIRVOL
The function boxvl.F is legacy code.  It returns the grid box
volume in cm3.  This can now be computed directly as

   BOXVL = State_Met%AIRVOL(I,J,L) * 1d6

so we can just replace references to BOXVL in the code with the
above line.   This should also make the code execute a little
more efficiently, as it saves on the overhead of a function call
for each grid box (I,J,L).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>