Eliminated triple-writing of NO-AN-$ diagnostic
Removed some calls to the DIAG2BPCH routine in diag3.F for
the ND32 "NO-AN-$" diagnostic section.  This corrected a problem
where this diagnostic was written out 3 times.

Also made some cosmetic changes to GeosCore/hcoi_gc_diagn_mod.F90,
mostly aligning columns to better F90 free-format.

NOTE: There seems to be a problem with the LNOX, as the SLBASE
array is returned as zero.  Still investigating.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>