Bug fix: Now compute CLDTOPS after reading in CMFMC
In GeosCore/geosfp_read_mod.F90, we had moved the CMFMC field
from the GEOS-FP A3dyn file to the A3mstE file.  But we did not
move the computation of CLDTOPS (which depends on CMFMC) to the
corresponding subroutine.  This is now fixed.

The symptom of this bug was that lightning NOx emissions were
zero for the 1st 3 hours of a GEOS-FP simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>