Prevent div-by-zero errors in HEMCO Rn-Pb-Be extension
In HEMCO/Extensions/hcox_gc_RnPbBe_mod.F90, the algorithm that was
used to compute F_BELOW_70 and F_BELOW_60 had a div-by-zero error
for the 2 x 2.5 grid for grid boxes on the equator.  The absolute
value of the latitude at the southern and northern edges were
subtracted from each other, which resulted in a zero that was then
placed in the denominator.  We have added a better error check that
prevents this situation from occurring.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>