Prevent div-by-zero error in routine GCAP_GLOBAL_ADV
We need to call SAFE_DIV to make sure the division of TR_DIFF by
SUMADA does not result in a div-by-zero.  If the division cannot
be done, then set TR_DIFF To zero.

The same error check has already been implemented for other met
fields, but not for GCAP.  This is now fixed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>