Added hacks to read proper O3, BC, OC filenames for GCAP
In order to allow GCAP POPs unit testing to proceed, add hacks in
global_bc_mod.F, global_o3_mod.F, and global_o3_mod.F to point
to the proper files on the GCAP grid.

NOTE: This will be a moot point once HEMCO is used to keep track
of these oxidant and concentration fields.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>