Update ND06 diagnostic for TOMAS simulations
The ND06 diagnostic is now compatible with the TOMAS dust tracers
(DUSTnn) as well as with the standard non-TOMAS dust tracers (DSTn)

NOTE: There seems to be a problem in CLEANUP_AEROSOL in deallocating
the SOILDUST array.  This may be an artifact of testing TOMAS w/o
chemistry turned on.  We hope that further unit testing will
reveal thes source of the error.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>