Add more flexible test for netCDF linker commands
Makefile_header.mk: Add a flexible test for the netCDF version that
replaces the hardwired test that we had previously implemented.
Also now use $(NC_INC_CMD) and $(NC_LINK_CMD) instead of older
variables $(NCI) and $(NCL).  $(NCL) is still exported for backwards
compatibility.

NcdfUtil/Makefile: Use $(NC_LINK_CMD) instead of $(NCL).

NOTE: NCL can also stand for "NCAR command language" so it is probably
a good idea not to have a variable named that in the Makefiles.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>