Restore changes in HEMCO that were clobbered by "Grande Merge" code
The "Grande Merge" updates (i.e. code that was made to fast-track
MEGA-Chem to the latest GEOS-5 tag "Heracles") clobbered some of the
code that was present in v10-01e-approved.  We have now restored
this code.

In particular,

(1) POA is now restored in FINN
(2) We now don't compute OCPI when MEGAN_SOA is turned on
(3) The bug fixes for the unit conversion in PARANOX are restored
(4) We now use the same hcoio_dataread_mod.F90 as v10-01e-approved

Unit tests w/r/t the v10-01e-approved version yielded identical results
for GEOS-5, MERRA, and GEOS-FP meteorology.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>