Declare time/date variables as 8-byte in additional modules
We have to make time/date variables declared as REAL*8 in many
extra subroutines.  Most of these are TAU0, TAU1 variables that
get placed in calls to BPCH2 or READ_BPCH2.

NOTE: In the near future we will be migrating away from bpch diagnostic
output.  Most of these diagnostics are going to be replaced by netCDF
output, as we move to the ESMF/MPI environment.  But we need to make
these changes in the code now in order to make it compile.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>