Bug fix: deactivate automatic diagnostics (AutoFill)
In GeosCore/hcoi_gc_diagn_mod.F90, a new IF block was created to turn on
automatic diagnostics.  This should just be used for testing, as it may
create problems with other simulations.  We have now added the following
modifications:

(1) Disable the autofill diagnostics by default

(2) Also wrap the autofill diagnostics in an IF block to make sure that
    they will only get executed if it's a full-chemistry simulation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>