Add fix for library link directory when using GEOS-Chem-Libraries
Added a fix to the Makefile_header.mk to override the library link
path returned by nf-config and nc-config if the $GC_LIB environment
variable contains the text "GEOS-Chem-Libraries".  This will make sure
that the code will find the proper library path.

If the GEOS-Chem-Libraries are built in one directory but then moved
to another directory, then the nf-config and nc-config commands will
return the directory where the libraries were built, but not where they
currently reside.  This fix addresses this issue.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>